updated compression analysis - accurate mass fractions must still be put in.
Suggestion for future-proof: the masses should be typed in here as a fraction of the actual mass of hermes II and multiplied by te mass from the sim-file name. As such we will hae the accurate masses for hermes analysis, and in the future we can just hermes mass fractions as a good place to estimate from.